!1090 Add metrics for hit ratio of ts block cache * fix: Fixed cpplint error. * fix: Fixed TestLint error. * fix: Fixed block cache unit test failure. * fix: Fixed TestLint errors. * fix: Fixed wrong initial miss count and changed memory size to 64 bits. * fix: Fixed status_test.go error. * fix: ts engine metrics * fix: Added ts engine metrics to fix compilation error. * Add metrics for hit ratio of ts block cache

+++ kwdbts2/engine/db.cpp
@@ -1360,7 +1360,7 @@ TSStatus TSFlushVGroups(TSEngine* engine) {
   return kTsSuccess;
 }
 
-void TsGetRecentBlockCacheInfo(uint32_t* hit_count, uint32_t* miss_count, uint32_t* memory_size) {
+void TsGetRecentBlockCacheInfo(uint32_t* hit_count, uint32_t* miss_count, uint64_t* memory_size) {
   TsLRUBlockCache::GetInstance().GetRecentHitInfo(hit_count, miss_count, memory_size);
 }
